Factor out common code between the two plugins
