+++ Plugins/AWSLambdaDeployer/Plugin.swift
@@ -79,7 +79,7 @@ struct AWSLambdaPackager: CommandPlugin {
         // Check if Deploy.swift exists. Stop when it does not exist.
         guard FileManager.default.fileExists(atPath: deploymentDescriptorFilePath) else {
             print("`Deploy.Swift` file not found in directory \(projectDirectory)")
-            throw PluginError.deployswiftDoesNotExist
+            throw DeployerPluginError.deployswiftDoesNotExist
         }
 
         do {
@@ -105,7 +105,7 @@ struct AWSLambdaPackager: CommandPlugin {
             let helperFileUrl = URL(fileURLWithPath: helperFilePath)
             try helperCmd.write(to: helperFileUrl, atomically: true, encoding: .utf8)
             try FileManager.default.setAttributes([.posixPermissions: 0o755], ofItemAtPath: helperFilePath)
-            let samDeploymentDescriptor = try execute(
+            let samDeploymentDescriptor = try Utils.execute(
                 executable: Path("/bin/bash"),
                 arguments: ["-c", helperFilePath],
                 customWorkingDirectory: projectDirectory,
@@ -124,14 +124,14 @@ struct AWSLambdaPackager: CommandPlugin {
                 to: samDeploymentDescriptorFileUrl, atomically: true, encoding: .utf8)
             verboseLogging ? print("\(samDeploymentDescriptorFilePath)") : nil
             
-        } catch let error as PluginError {
+        } catch let error as DeployerPluginError {
             print("Error while compiling Deploy.swift")
             print(error)
             print("Run the deploy plugin again with --verbose argument to receive more details.")
-            throw PluginError.error(error)
+            throw DeployerPluginError.error(error)
         } catch {
             print("Unexpected error : \(error)")
-            throw PluginError.error(error)
+            throw DeployerPluginError.error(error)
         }
         
     }
@@ -144,7 +144,7 @@ struct AWSLambdaPackager: CommandPlugin {
         guard let executable = try? context.tool(named: executableName) else {
             print("Can not find `\(executableName)` executable.")
             print(helpMessage)
-            throw PluginError.toolNotFound(executableName)
+            throw DeployerPluginError.toolNotFound(executableName)
         }
         
         if verboseLogging {
@@ -164,21 +164,21 @@ struct AWSLambdaPackager: CommandPlugin {
         print("-------------------------------------------------------------------------")
         
         do {
-            try execute(
+            try Utils.execute(
                 executable: samExecutablePath,
                 arguments: ["validate",
                             "-t", samDeploymentDescriptorFilePath,
                             "--lint"],
                 logLevel: verboseLogging ? .debug : .silent)
             
-        } catch let error as PluginError {
+        } catch let error as DeployerPluginError {
             print("Error while validating the SAM template.")
             print(error)
             print("Run the deploy plugin again with --verbose argument to receive more details.")
-            throw PluginError.error(error)
+            throw DeployerPluginError.error(error)
         } catch {
             print("Unexpected error : \(error)")
-            throw PluginError.error(error)
+            throw DeployerPluginError.error(error)
         }
     }
     
@@ -196,7 +196,7 @@ struct AWSLambdaPackager: CommandPlugin {
         print("-------------------------------------------------------------------------")
         do {
 
-            try execute(
+            try Utils.execute(
                 executable: samExecutablePath,
                 arguments: ["deploy",
                             "-t", samDeploymentDescriptorFilePath,
@@ -204,14 +204,14 @@ struct AWSLambdaPackager: CommandPlugin {
                             "--capabilities", "CAPABILITY_IAM",
                             "--resolve-s3"],
                 logLevel: verboseLogging ? .debug : .silent)
-        } catch let error as PluginError {
+        } catch let error as DeployerPluginError {
             print("Error while deploying the SAM template.")
             print(error)
             print("Run the deploy plugin again with --verbose argument to receive more details.")
-            throw PluginError.error(error)
+            throw DeployerPluginError.error(error)
         } catch {
             print("Unexpected error : \(error)")
-            throw PluginError.error(error)
+            throw DeployerPluginError.error(error)
         }
     }
 }
@@ -251,7 +251,7 @@ private struct Configuration: CustomStringConvertible {
             guard
                 let buildConfiguration = PackageManager.BuildConfiguration(rawValue: buildConfigurationName)
             else {
-                throw PluginError.invalidArgument(
+                throw DeployerPluginError.invalidArgument(
                     "invalid build configuration named '\(buildConfigurationName)'")
             }
             self.buildConfiguration = buildConfiguration
@@ -267,7 +267,7 @@ private struct Configuration: CustomStringConvertible {
         }
         var isDirectory: ObjCBool = false
         guard FileManager.default.fileExists(atPath: self.archiveDirectory, isDirectory: &isDirectory), isDirectory.boolValue else {
-            throw PluginError.invalidArgument(
+            throw DeployerPluginError.invalidArgument(
                 "invalid archive directory: \(self.archiveDirectory)\nthe directory does not exists")
         }
         
@@ -302,9 +302,8 @@ private struct Configuration: CustomStringConvertible {
     }
 }
 
-private enum PluginError: Error, CustomStringConvertible {
+private enum DeployerPluginError: Error, CustomStringConvertible {
     case invalidArgument(String)
-    case processFailed([String], Int32)
     case toolNotFound(String)
     case deployswiftDoesNotExist
     case error(Error)
@@ -313,8 +312,6 @@ private enum PluginError: Error, CustomStringConvertible {
         switch self {
         case .invalidArgument(let description):
             return description
-        case .processFailed(let command, let code):
-            return "\(command.joined(separator: " ")) failed with exit code \(code)"
         case .toolNotFound(let tool):
             return tool
         case .deployswiftDoesNotExist:
@@ -325,116 +322,3 @@ private enum PluginError: Error, CustomStringConvertible {
     }
 }
 
-// **************************************************************
-// Below this line, the code is copied from the archiver plugin
-// **************************************************************
-@discardableResult
-private func execute(
-    executable: Path,
-    arguments: [String],
-    customWorkingDirectory: Path? = .none,
-    logLevel: ProcessLogLevel
-) throws -> String {
-    if logLevel >= .debug {
-        print("\(executable.string) \(arguments.joined(separator: " "))")
-    }
-    
-    var output = ""
-    let outputSync = DispatchGroup()
-    let outputQueue = DispatchQueue(label: "AWSLambdaPackager.output")
-    let outputHandler = { (data: Data?) in
-        dispatchPrecondition(condition: .onQueue(outputQueue))
-        
-        outputSync.enter()
-        defer { outputSync.leave() }
-        
-        guard
-            let _output = data.flatMap({
-                String(data: $0, encoding: .utf8)?.trimmingCharacters(in: CharacterSet(["\n"]))
-            }), !_output.isEmpty
-        else {
-            return
-        }
-        
-        output += _output + "\n"
-        
-        switch logLevel {
-        case .silent:
-            break
-        case .debug(let outputIndent), .output(let outputIndent):
-            print(String(repeating: " ", count: outputIndent), terminator: "")
-            print(_output)
-            fflush(stdout)
-        }
-    }
-    
-    let pipe = Pipe()
-    pipe.fileHandleForReading.readabilityHandler = { fileHandle in
-        outputQueue.async { outputHandler(fileHandle.availableData) }
-    }
-    
-    let process = Process()
-    process.standardOutput = pipe
-    process.standardError = pipe
-    process.executableURL = URL(fileURLWithPath: executable.string)
-    process.arguments = arguments
-    if let workingDirectory = customWorkingDirectory {
-        process.currentDirectoryURL = URL(fileURLWithPath: workingDirectory.string)
-    }
-    process.terminationHandler = { _ in
-        outputQueue.async {
-            outputHandler(try? pipe.fileHandleForReading.readToEnd())
-        }
-    }
-    
-    try process.run()
-    process.waitUntilExit()
-    
-    // wait for output to be full processed
-    outputSync.wait()
-    
-    if process.terminationStatus != 0 {
-        // print output on failure and if not already printed
-        if logLevel < .output {
-            print(output)
-            fflush(stdout)
-        }
-        throw PluginError.processFailed([executable.string] + arguments, process.terminationStatus)
-    }
-    
-    return output
-}
-
-private enum ProcessLogLevel: Comparable {
-    case silent
-    case output(outputIndent: Int)
-    case debug(outputIndent: Int)
-    
-    var naturalOrder: Int {
-        switch self {
-        case .silent:
-            return 0
-        case .output:
-            return 1
-        case .debug:
-            return 2
-        }
-    }
-    
-    static var output: Self {
-        .output(outputIndent: 2)
-    }
-    
-    static var debug: Self {
-        .debug(outputIndent: 2)
-    }
-    
-    static func < (lhs: ProcessLogLevel, rhs: ProcessLogLevel) -> Bool {
-        lhs.naturalOrder < rhs.naturalOrder
-    }
-}
-
-
-// **************************************************************
-// end copied code
-// **************************************************************
